ARM: Improve switching to HYP mode
Make our approach to switch to HYP mode more user
friendly: Tell the user that the CPU does
not support HYP mode and thus switch to HYP
mode cannot work. This happens when running
a HYP-enabled kernel in QEMU. Also tell
the user that the currently used platform
does not provide switching code and it thus
cannot work.

Change-Id: I825f32dd16fad885ecfb4f94ec87f15d553eefda