Avoid make pattern rules rebuild modules
To figure out whether we need to regenerate the
bootstrap image, we used a make rule.
However, this has the effect that make
tries to apply its rules on the modules file
itself, e.g. when having a dtb-file in the modules.list,
the %.dtb rule kicks in and tries to rebuild the
dtb-file, and fails.
For this reason we explicitly list the files to
generate targets, avoiding that the pattern
rules kick in.

Change-Id: I04e2814350f8fe716859f5683954da9b80d9b483