x86/amd64: Enable FPU for SSE2
Code used with and by bootstrap might contain
SSE2 code as on 64-bit SSE2 is always available.
However, for that the FPU must be enabled.
To be able to use that code (and not require a third
variant of the libraries), we enable the FPU for
bootstrap and reset it again when booting Fiasco.

Change-Id: Ide9ba4c7d30a37acfc902445569264f6209ee2a7