Add support for multiboot2
This change enables multiboot2 boots. With many other parts of the
system depending on multiboot (v1) data structures and with the goal of
making a low-impact change, the method chosen is to extract essential
information from the multiboot2 info structure and convert it into a v1
structure as soon as possible so that the rest of the bootstrap code
remains untouched.

Fixes ticket #526

Change-Id: I25d1ac76d17d9bbc70702e876dc2774feb368d0c
Requires-Change-Id: I753271defdfaa924cda6dbba3aa686bfe4102c52