x86: move conventional EBDA search
On modern systems (esp. UEFI with CSM) the conventional search for the
EBDA may produce erroneous results. The BDA is obsolete nowadays and
reading a value from 0x40e (EBDA base address) may produce a false
positive as there might be other data laying around. In the end this may
confuse bootstrap to reserve a memory area that it later needs to load
the L4Re microkernel.

Instead we should only rely on the platform memory map to reserve memory
areas. Thus we move the conventional EBDA search to the code path where
the bootloader does not provide us with a valid memory map.

Change-Id: I0246dc811b29da2351f28e331f293a378c5b2e2a