x86: reserve address range type 20 memory
On newer x86 machines I observed address range type 20 memory in the
memory map. According to the ACPI specification this is an undefined
type reserved for future use. However we must treat this range as
reserved memory.

This change checks for ART 20 ranges and reserves the region as
architecture specific.

Change-Id: I483af3fc72ef63f1fedffcf381323b6fafce84f6