amd64/boot_cpu: avoid warning
With GCC 8.2 we get "warning: unsupported size for integer register"
when trying to set %cr0 to enable paging. This is due to the compiler
expanding (get_cr0() | const_val) to 64 bit because the enum has
larger-than-int types. Fix this by splitting the enum so
that the type in question (CR0_PG) is 32bit.

Change-Id: Ied9d69d6017cd847d98a355a1ca498f6f52f07e2