Fix parallel building
Some target files for bootstrap are created in a subdirectory which has
to be created first. For parallel builds this is racy and occasionally
leads to build failures. The reason for that seems to be the use of
'eval' within a make rule.

This change removes the usage of 'eval' and instead invokes function
that does the right thing and eventually creates the directory if
necessary.

This fixes the github issue as reported by @skalk.

Change-Id: I1d7c03fdcc6497b13171af4b82b47639c57cec66