Refine PLATFORM-existence check
Certain versions of nm (like the very recent ones of v2.32+) show
the checked symbols as 'B' instead of 'D'.
This looks like a wrong display of nm, objdump shows it
correctly, however, we need to cope with this and thus
relax the search pattern to any letter, as we do not
particularly care for the specific letter.

Change-Id: I2e99070a562950293be0c66a9656b0dfb3507844