Fix handling of fname
The tooling in mk's tool/bin has various little inconsistencies in the
handling of the fname feature which happens due the fact that these
tools either ignore what is specified in the cmdline key or misuse
this information leading to missing modules, wrongly renamed modules
or outright wrong filenames in bootloader configurations.

To fix this we extend the $mod structure with more information, such
as what the module is called on the build machine an what it should be
called in the environment of L4Re.

Beside that there are various instances extracting arguments from the
command line using regexes. Instead of doing this we add the arguments
as another key within the mod structure so that the arguments can
easily be referenced.

The mod structure now contains the following keys:

- file: file name/path as specified in the modules.list
- command: file name of the module for the L4Re environment
           (fname if specified otherwise basename of the file key)
- cmdline: command + ' ' + args if args are specified, otherwise just
           command
- args:    Arguments as specified in modules.list
- cmdline_quoted: cmdline but " replaced with \"
- args_quoted: args but " replaced with \"
- opts: options hash

Change-Id: I2fe8b8997ce721d3353da96d7a19399e677e022d