x86: disable PCI bus mastering
System firmware may enable an IOMMU prior to iterating and configuring
the PCI bus to protect itself against malicious DMA. Before handover to
an OS it has to turn the IOMMU off because system firmware does not know
whether the OS is IOMMU aware. This opens a small window of opportunity
for malicious PCI devices.

If we trust the chipsets built-in PCI bridges to function properly we
are able to close this window of opportunity by disabling PCI bus
mastering.

This change disables PCI bus mastering on all found PCI devices.

Change-Id: Id14739af2242a5f58a8c6fe5178d22fdfdd66b68