Make.rules: fix dependency tracking of OBJ32 objects
The build rules for OBJ32 objects are hand-woven in Make.rules because
they build for the i386 target instead of x86_64 as the rest of
bootstrap. These rules used the wrong flags to create dependency files
as byproduct. Additionally the silent include of these dependency files
was missing too. As a result, changes in transitive dependencies did not
lead to a rebuild of these special *.o32 objects.

This bug did only affect x86_64 builds because the special OBJ32
handling is only used there.

Change-Id: I3a0c27fd488f2ee5611cc07277d38f2085cff290