improve PCI serial card discovery
The current discovery code to search for PCI serial cards only matches
for device and vendor id to find a driver. The driver then initializes
the card by calling the setup() method.

For cards with the MOS chip this would also initialize the parallel
port.

To fix that the setup() method of the driver checks for the correct PCI
class and subclass to make sure it is actually a serial port.

Change-Id: I669f98b0b6fe36fa64e9095a5d81029785e09d54