modules: Make module uncompression explicit
Up to now we have implicitly uncompressed every
module. This is counterproductive for cases where
the module shall be compressed, e.g. for a compressed
ramdisk/initrd which guests can unpack themselves.
For this reason, we only uncompress when requested
via the new 'uncompress' option.
Uncompressing is still useful, for example, to save
space on the local disk.

Change-Id: Id1455b1cce3533cbd187c253ac5f7e5281bb6a9b
Requires-Change-Id: I2fb0c5c99cb899cf50f8753aa2720e2247a1ef7b
Requires-Change-Id: I80565199427e7266f3f47ab2c956c6beaec12f0b