optimize memcpy() for 8-byte aligned memory
Bootstrap uses memcpy() to copy the system modules (Sigma0, root task
and the kernel) to the intended addresses. For these ELF binaries we
can assume a section alignment of at least 8-bytes. Use an optimized
memcpy() implementation working on aligned memory which increases the
throughput on certain architectures (e.g. ARM64/RCar3) by a factor of
4-5. There should be no fundamental drawback on other architectures.

Fall back to the LibC-provided memcpy() implementation if either the
source address or the target address is not properly aligned.

This optimization is only useful in the bootstrap context (caching
disabled, no MMU). Executing the same code in L4Re userland does not
show any improvement comparing to the LibC memcpy() implementation.

Some values for RCar H3 (Cortext A57) -- lower values are better:
* bootstrap / memcpy():              37,933us/MiB
* bootstrap / memcpy_aligned():       7,755us/MiB
* L4Re userland / memcpy():          10,220us/MiB
* L4Re userland / memcpy_aligned():  10,249us/MiB

On this platform it really takes at least 7.7ms to copy one MiB.

Change-Id: I4081cd2110e9b77ab843687edfbd6e6852f53009