amd64/elf: Put inner ELF in an extra data section
On AMD64 systems bootstrap is compiled as a bare bones 32 bit variant
that sets up long mode and then loads a 64 bit variant. That 64 bit
variant is an extra ELF inside the 32 bit binary.

To use l4image with such netsted ELF binaries it must be able to
identify the inner image and be able to replace it if necessary. This
change modifies the linker script such that the inner ELF is put inside
a .data.elf64 section. To prevent the linker from merging those sections
separate phdrs have to be specified.

Change-Id: I0c7bedbfe2db3d1a22abcd741860bc1fba0101c2