boot32: make overlap checks more flexible
The overlap check currently only checks if the target memory area
overlaps with the image itself. This changes makes the checking more
flexible by having a map that is filled at runtime with regions that
must not overlap the image.

This will allow future changes to also protect modules and MBI data
structures from being overwritten.

For now this only converts the current mechanism to a new behavior.

Change-Id: I57bedca3e3a6e4312d0c479ed687142fde8fa9ca