boot32: Relocation resolution for outer bootstrap
Enable the outer bootstrap to load a position-independent inner (64bit)
bootstrap image. It also adds a very simple allocator that works on the
MBI memory map to find suitable memory locations to load an image to
while preserving structures required in later boot stages.

To enable the feature, build the bootstrap image with BUILD_PIE=1 set in
the environment.

Change-Id: I8dfc94c052afee59bada73719e7d1ef6c1ced912
Requires-Change-Id: I79b1d78928dd5016a66047ecf6694e203b01b842