boot32: remove unused check and protect MBI data
The old check in check_mbi_modules_in_limit was making sure that modules
would not lie beyond the memory mapped in the page table used for the
inner bootstrap. Since this limit was hard-coded to 4GiB, effectively
all modules would lie within the limit as the MBI structure for modules
only works with 32bit values.

Being able to limit the page-table size and thus the available memory
for the inner bootstrap is no longer required on current memory-rich
amd64 systems. Thus the check is removed.

The commit also protects data structures of the MBI from being
overwritten by the inner bootstrap.

Change-Id: I85db1cb350c43d2b624cf6b9b49d0b114921eaad