find_kip: Fix start of KIP scanning range
When scanning for the KIP the search should not round down the start
address of a region since that would mean starting the search before the
region start. Instead use the rounding function from l4sys.

Also improve general code legibility by using a C++ for-each loop to
iterate over regions and simplify the remaining code.

Change-Id: I72248d3e0652684702614dcbdcdc4c11713992db