elf loader: Improve legibility through inheritance
At the moment the elf_loader_exec assumes that the pointer passed to it
can be interpreted as a pointer to the address of an ELF binary. Callers
exploit this by passing in structs that have a boot module as first
member and additional information that can be passed to the exec
handlers.

Instead of having magic knowledge about the structs fed into the
function provide a class that has a boot module as first member and let
the other structures inherit from it. Also remove a const_cast that is
now no longer needed.

Change-Id: I6a6e3155944a3d416a25026bc08c7a96b6320848
Requires-Change-Id: I89d8844ca6d8818f4b8e652ab7a93ab45b40e4b7