platform: Add initial set of Qualcomm Snapdragon platforms
These are used in various popular smartphones, tablets, smartwatches,
laptops, embedded/automotive boards, ...

- Qualcomm Snapdragon 400 (MSM8226)
- Qualcomm Snapdragon 210 (MSM8909)
- Qualcomm Snapdragon 410/412 (MSM8916)
- Qualcomm Snapdragon 615 (MSM8939)
- Qualcomm Snapdragon 800 (MSM8974)
- Qualcomm Snapdragon 855 (SM8150)

This initial set just corresponds to those that have been tested
successfully on some device so far, more platforms could likely
be added easily with reasonable effort.

Unfortunately it ends up being a quite confusing maze of #if macros to
handle address differences between all the platforms. There is no good
way around that, a similar approach is already used in imx.cc for
example.

In the future it might be possible to also read the UART address from
the device tree instead. However, this would require some rework in
bootstrap to parse the device tree much earlier. The disadvantage is
that it would be no longer possible to have debug console output when
the DT is missing or invalid.

Change-Id: Ibd06520187afdad4aa5b924512c56230e7a4dff9
Requires-Change-Id: I795f2d223af558b70c10e098ef7e1b9a5b8aa11a
Requires-Change-Id: I6b8d8b47e66676d8f8ff276809b271453af4d256