introduce utest feature string replacement
This commit introduces the new command line option '-utest_opts' for
the kernel. It takes a string encoded configuration for the kernel
unit test framework as parameter.
If the command line option is given, the code searches for the kernel
feature string prefixed with 'utest_opts='.
This feature contains a placeholder string after the prefix, which
is replaced with the configuration string.

Change-Id: I3d169f408f20aa9b42be7df753fe64d18827c5b2