arm64: Mimic ARM64 vmlinux image similar to ARM32
Some bootloaders insist on having the Linux kernel header at the
beginning of all booted images, e.g. to decide if the image should be
booted in ARM32 or ARM64 mode. In bootstrap the functionality to "mimic"
such a vmlinuz header already exists for ARM32 but not for ARM64.

Add it for ARM64 as well similar to the ARM32 implementation. Platforms
that do not need this will just skip over this code because it is
effectively a NOP.

For simplicity and best compatibility the header is filled like on older
Linux versions (prior to v3.17). There image_size was always 0 and
text_offset was always 0x80000 [1].

[1]: https://www.kernel.org/doc/html/latest/arm64/booting.html#call-the-kernel-image

Change-Id: I3d3e45b397b146545a22cb6d51895be68077932e