x86: change confusing init of ofs
Change initialization of ofs. First, the value is strange becase ofs is
not a memory offset and thus does not need the '2' in there. Second,
it's not relevant, because vga_putchar('\n') will set it to 0 anyway. So
prevent anyone else thinking about this by setting it to 0 right away.

Change-Id: Ib36eacf0e5b56598b3fde2418d223ac050ae1bf6