amd64: Fix XHCI handoff for 64-bit BARs
Bootstrap tries to do XHCI handoff with the BIOS early
in order to avoid issues with the IOMMU when it is later
turned on by fiasco.

However, this code does not handle 64-bit BARs correctly.
Thus handoff takes somewhat long, accesses arbitrary memory
locations and in the end still does not work.

If we fix this it turns out that the 32-bit bootstrap
code did not setup identity mappings for the high
BAR location above 4GB.

To fix the latter issue, during XHCI handoff we ensure that the
first 256 bytes (maximum size of the XHCI capability registers)
are identity mapped, if necessary we create this mapping.

Since the creation of this mapping might require the allocation
of page tables, the XHCI handoff is now executed at a later point
in time during the startup than before. Namely, when the memory map
is set up and all modules have been loaded or moved.

This bugfix is adapted from a submission by:
Christian Ehrhardt <ehrhardt@genua.de>

Change-Id: I018d1263ad9912d4b5c741751ece03f957d059bd