Add support for AMP multi node boot
AMP is modeled around the concept of "nodes". Each node (read: non
cache coherent CPU), runs a separate instance of L4Re. The difference
to a purely heterogeneous system is that some parts are still shared.
Namely the kernel image, the modules and access to RAM. Obviously, the
sigma0 and root task applications need to coordinate themselves between
the nodes.

For AMP, the kernel provides a dedicated KIP and koptions for each
node. Bootstrap needs to match the right KIP/koptions and fill in the
relevant data for that node. Which memory regions are applicable to
what node is encoded as a bitmask in the module information and the
bootstrap regions.

Modules may use the "attr:nodes=..." attribute to specify the
applicable AMP node of sigma0 and the root task. It takes a colon
separated list of ranges, e.g. "attr:nodes=0:2-3". If unspecified, it
defaults to all AMP nodes.

Requires-Change-Id: Ie65e6498e267c02545162ec0a22185f964c6e246
Change-Id: I1d2f78bee0350396065b3091d4336335010505c9