ensure misaligned conventional memory is used 1K-aligned
A memory region with an non-1K-aligned start or end address is expanded
to the next larger 1K-aligned region when the start and end address are
passed to Kip::Mem_desc's constructor.
This behavior is wrong for conventional memory types, as the physical
memory cannot be expanded.

In case of conventional memory, the next smaller 1K-aligned region must
be passed to Kip::Mem_desc.

Change-Id: I74a96e1d57387caf3a677b44512ed72cfc193829