allow Boot_modules::base_mod_idx() to fail
The function Boot_modules::base_mod_idx() is supposed to return the
index of a certain boot module (kernel, sigma0 or roottask). Allow that
this function fails if the corresponding module is not available and
return -1 in that case.

Remove the functions Boot_modules::mod_{kern,sigma0,roottask}() and
store the corresponding module indices instead (the index could also be
-1 if the corresponding module is not available).

Change-Id: I95b60357831c64838efd0431f1e6ef9995c179de