amd64: Prevent .data.elf64 section from being garbage collected
The linker invokation for bootstrap32.elf sets the --gc-sections flag.
If for some reason the compiler determines that the code referencing
_binary_bootstrap64_bin_start will never be executed, it optimises
away this single reference to the .data.elf64 section, whereupon the
linker determines that the section is unused and therefore can be
garbage collected. This then results in an empty rodata PT_LOAD phdr.

And unfortunately it seems that the QEMU elf loader / firmware cannot
handle empty phdrs, i.e. it crashes while loading such a bootstrap32.

It may seem unlikely that gcc optimises away the reference to
_binary_bootstrap64_bin_start, thus triggering this chain of problems.
But actually this can happen quite easily, e.g. if you insert an endless
loop or a panic somewhere in bootstrap32 for debugging purposes such as
to see the printfs of bootstrap32 on the VGA console.

Change-Id: I7a841405cac37c5b19654a8bf5a628ee293dc2c0