Interpret the UEFI load options as the command line
The UEFI load options are useful to pass the custom UART configuration
to bootstrap (and potentially other options in the future) from the
platform boot loader or from the UEFI boot menu.

This also fixes the problem where the mod_info_mbi_cmdline(mod_header)
was called before the call to init_modules_infos(), thus with the
mod_header structure potentially uninitialized.

Change-Id: Id3df4af76fa3be988b2406123bd774c22a97e131