x86/EFI: Look for and setup EFI GOP (Framebuffer)
This sets up the EFI GOP (framebuffer) and stores it in the MBI video
infno fields, such that the framebuffer is setup and can be used by
other applications.
Note that if the reported framebuffer address is beyond 4G it will be
rejected because the current MBI structure can only hold 32 bits. It's a
todo to rework this structure.

Change-Id: If6bf2c6ca02a9b16b4e45332948cc7f0a4aa28c2