Add support for QEMU ramfb
QEMU ramfb is a simple framebuffer display device, which is intended to
be configured and used by the guest during boot, until the guest
operating system is ready to switch to the real GPU. The guest allocates
memory for the framebuffer from guest RAM and then initialises the ramfb
device via the QEMU firmware configuration interface.

Bootstrap sets up the ramfb device, if present, then populates the MBI
video info fields, so that the framebuffer can be discovered and used by
subsequent applications.

Change-Id: I8fa4a90eb7cf30419daea8108415da7fdd519249