Arm: Take device tree address from r2 on ARM32
Most bootloaders such as U-Boot pass the device tree pointer in the r2
register when booting ARM32 kernels. This is because most of them are
originally designed for Linux and the ARM Linux boot protocol expects it
in r2 and not r0 [1].

Perhaps this did not cause any problems so far because all the currently
supported DT platforms also support ARM64 where it is indeed passed in
x0 [2]. QEMU also passes the DT pointer in r2 on ARM32 but is not
affected only because the DT address is currently hardcoded in
Platform_arm_virt.

[1]: https://www.kernel.org/doc/html/latest/arm/booting.html#calling-the-kernel-image
[2]: https://www.kernel.org/doc/html/latest/arm64/booting.html#call-the-kernel-image

Change-Id: Iec0222ff9b79ce919313b6b05a4d8aa4d34f41a5