x86: Fix BAR mapping during xHCI handoff
Bootstrap incorrectly assumed that all xHCI extended capabilities are
located in the "Capability Registers" register set, which is located at
offset 0 of the xHCI memory-mapped register space and has a limited
size. Therefore, only the first page of BAR0 was mapped. In fact,
however, this is not correct, and the extended capabilities can in be
anywhere in the memory space provided by BAR0.

If the extended capabilities were indeed outside the first page, this
resulted in the machine getting stuck during startup in bootstrap.

With this change bootstrap now maps the entire BAR and in addition
checks that the extended capabilities lie within the BAR's memory space.

Change-Id: Ib2f1014cbef8a39a24f1aec8b09364ce34c73ec4