build.pl - Add attr if image is for uefi or not
When using the l4image tool to build tests, we need to make sure we use
the right base image together with l4image tool to create the test
image. For this we have to check a few properties of the base image: The
selected platform, the ram base as well as whether or not the image was
built as an EFI image. This change adds the latter information as
attribute to the image.

Change-Id: I107115158438f8bebcca8e154af3395b1f5445c2