Implement support for efi images in l4image
This change implements support for EFI images in the l4image tool. EFI
images are PE binaries with a special PE32+ sub header. To add support
for them we add another L4::Image module called L4::Image::EFI that can
read and adjust these headers. Due to the structure of the PE binary
dictated by gnu-efi each of our efi images has two sections, one text
section and one data section, with the data section being the second
one. The L4::Image::EFI module adjusts the size of the data section
according to how the module data grows and shrinks. However for this to
work the module data must sit at the end of the PE data section, so that
no other data is pushed around by growing or shrinking the module data.

Change-Id: I1cde1fcd47bfe93b8a076efa7f64ad78b4398033