startup - Ignore initial mbi regions during overlap check
On x86_* we temporarily create .mbi regions that are discarded again as
soon as we gathered the data inside a final structure. We need these
regions so that our final structure is not allocated in the same space
as the .mbi regions. However, we also need this final structure to be
outside of the regions that fiasco, sigma0 and moe will inhabit in the
future, which is why we create these regions beforehand. The problem is
that these initial .mbi regions might collide with the fiasco, sigma0
and moe regions although we will actually not use the latter before a
much later point in time. To fix this, we mark the .mbi regions with a
special subtype and ignore them during the overlap check for the fiasco,
sigma0 and moe regions.

See ticket #CD-1051

Change-Id: I0bec93bd7bb45442621e5fd3ced30c5a4ef4d3d2