Make memory preseting work reliably with large amount of RAM
Since the bootstrap might not have access to the entire physical memory
range on some platforms/architectures (e.g. on AMD64), using the
-presetmem= option on machines with a large amount of RAM could lead to
a crash.

To address this, we introduce a new global variable mem_end that can be
set by the platform/architecture-specific code to limit the maximum
virtual address that bootstrap should access. Without an explicit
override, the default value leads to the previous sensible behavior.

The fill_mem() function has been also refactored to simplify the
algorithm and make it more straightforward. The implementation now uses
the same (inclusive) semantics for the end addresses as the surrounding
code. This, among others, avoids the processing (and a slightly
confusing printout) of zero-sized ranges. The comments have been
extended.

Finally, the debugging printout in bootstrap() for AMD64 has been
altered to express the semantics of the value better.

Change-Id: Id3d4c02ec8df4d67743634c89b7cb6d36cfaad7b