x86: cosmetic changes for scanning PCI UARTs
First, don't list all scanned PCI devices but only devices which are
detected as UARTs. Listing all PCI devices does not make much sense
because the space on the VGA screen is limited and most of the listed
devices scroll out of the screen anyway.

Second, print messages when starting the scan and when the scan has
ended.

Also, don't pretend to call reboot after scanning the PCI bus. Here we
use the default implementation of Platform_base::reboot() which just
executes the infinite loop. Hence, print an appropriate message and call
l4_infinite_loop() directly.

Change-Id: Idd0e9d294d86737e1ad63c5d90680802f72a95b4