boot32: avoid linking against host libgcc
The host libgcc variant might be built to use AVX instructions, which
are not available in a bootstrap context. Do not link against it but
provide a simple implemnetation of our own.

This fixes an illegal instruction trap in boot32 when building on a
Gentoo system where the libgcc is tailored towards the host CPU.

Change-Id: Ia841b354ba72e595d3dc252d6e4fc1eac34d270e