arm: Remove virtualization configuration warning
Since I2b949d873d541dbe16ebb124e53a258903030dcc the userland runs in EL0
again and can use the same SVC instruction irrespective of if the kernel
runs in EL1 or EL2. Thus remove the old misconfiguration warning.

This was also the only user of the module_load_hook function, so remove
that as well.

Change-Id: I333169191971e7f6591be29cc5009417c9598640