x86: make sure image info is not dropped
The Image_info structure must always be present in the binary because
it is required by the l4image tooling. On x86 (or multiboot to be
precise) this is only accidentally the fact because some symbols in the
object file are referenced. Make the reference explicit instead to
prevent the linker from ever dropping it.

Change-Id: Ica957246c5269b22cf98e6765b3b32ab4375a052