fix global/weak conflict with platform_cpu_bootup
ARCH-arm/crt0.S and ARCH-arm64/crt0.S invoke platform_cpu_bootup().
Default implementations of this functions are implemented in the same
files. Hence, the platform_cpu_bootup() function label in these files
should not be "global" but should carry the "weak" attribute.

In contrast, platform-specific implementations of this function in other
modules should be ".global" but not "weak".

This fixes the following Clang warning:

  /tmp/user/1000/crt0-0865f7.s:106:1: warning:
    platform_cpu_bootup changed binding to STB_WEAK
    .weak platform_cpu_bootup

Change-Id: I3e8ab48bba8fe94dd2c8672c21359c503482d507