[API] Move module compression to l4image
This change enables module compression support using gzip to all
l4image commands where it is deemed useful. Editing or adding files,
extracting and repacking an image respect the flag and set it in
generated modules list files.

The new `compress` flag can be specifed per-module or by using the
compress-all option of l4image to compress all modules for an entry. If
a module is requersted to be decompressed by l4image and then compressed
again, a warning will be issued.

Change-Id: Ia5a55609f5fffcf05e08fcd59d7c674dcf926be6