Add Kconfig option and feature flag for decompression
This makes the option better visible and decouples it from the forced
compression of all modules. Beware that only specifying COMPRESS now on
the command line will not enable decompression in bootstrap as well.

A feature flag attribute indicates the capability for decompression in
bootstrap.

A warning will be issued if modules are compressed but decompression is
not built in.

Change-Id: I1d4d12e778f2ed9fe92d3cc60df325cdef42d5f4