arm: drop crt0 spin address boot code for qcom
The Qualcomm platform port supports and uses the spin address boot
protocol for older platforms without PSCI support, but only through the
"cpu-release-addr" property set by the bootloader in the device tree. In
this case, the CPU spins in a region initialized and reserved by the
bootloader. The code that was made optional in change
I0c8b38fe08642908abc0d490af2ac8b8bfd9317c is only needed when all CPUs
enter the bootstrap entry point and the CPUs are set up to spin inside
bootstrap code. This should never happen here.

The "cpu-release-addr" property is checked unconditionally in
server/src/platform_dt-arm.h, so the extra code is not needed for qcom.

Change-Id: I594f93c645135938fb070c2fe404206e8303a821