arm_sbsa: Allow skipping serial UART initialization
Check for the -noserial parameter given to bootstrap to skip
initialization of the serial UART. This can be used if printing to the
serial console is not desired or as a workaround if the information in
the SPCR ACPI table is incorrect.

Also move the check for the supported UART types so that the panic is
skipped if an unsupported UART is used together with -noserial.

Change-Id: Id0b3825c1ca1326f4f750b25ebd403ac7d5035c3