crt0: fix PIE relocation in thumb mode
If CONFIG_BOOTSTRAP_THUMB_ENTRY is enabled, the entry point (_start) is
a thumb mode symbol. This causes the LSB of the respective GOT entry to
be set to denote the thumb mode. We have to clear this bit when
calculating the relocation offset. Otherwise the result will be off by
one, leading to boot failures in this configuration.

Change-Id: Id153fa5d8ca5310db2d7a8950b6d082b96234c3b