Add "reloc" module attribute
Normally the sigma0 or root task images are loaded at the preferred
load address if the RAM is available at the desired location. But on
no-MMU systems it might be desirable to always relocate PIE binaries.
There we want to maximize the size of contiguous free RAM regions.
Without an MMU the fragmentation in the physical memory space cannot be
hidden.

Setting the "reloc" module attribute to a non-empty string will now
always request the dynamic relocation.

Change-Id: I0c063b99bb888336b89fc69c5600f4f40b9a9261