amd64: Also install a bootstrap_NAME.elf32
When building an ELF image for amd64, the bootstrap_NAME.elf is a
32bit binary which is the way to boot. However, when building an EFI
image, the bootstrap_NAME.elf is a 64-bit binary and actually an
intermediate file for building the final EFI image. To keep the working
32bit ELF file available, we make that available as
bootstrap_NAME.elf32. Thus use bootstrap_NAME.elf32 or
bootstrap_NAME.efi as required on amd64.

Change-Id: Ia5e75fbb46c521ec73848e83bfc4a3a5aaebe6f7