DT: Decouple init and module registration
Decoupling the initialization of the device tree and the registration of
the device tree module allows the former to happen before we have the
module structures. In an upcoming change, information from the device
tree will be used to setup the UART, so we want to have the device tree
available as early as possible.

Change-Id: I322c12bdd9d23146080de63099782303a289845a