startup: do not reserve memory for inactive AMP nodes
The kernel has a KIP in its image for each supported AMP node.
Bootstrap will only load sigma0 and the root task for nodes that are
supported. Unfortunately, we unconditionally reserved the memory for
such nodes which wasted precious memory. Remove this oddity by checking
for the node support already in the reservation stage.

Change-Id: Ia0fb9b3c45d02321e6d9f0f3760fd7dd8eb19760