Align sections to accommodate some strict boot loaders
Some strict boot loaders (e.g. the kexec mechanism) have requirements on
the alignment of the loaded sections (or possibly on the alignment of
the related ELF Program Header entries) that go beyond the multiboot
specification. Sadly, without respecting the alignment requirements,
they refuse to boot the image.

We accomodate these strict boot loaders by aligning the loaded sections
to a page boundary. While this might render the boot image slightly
larger, it should have no noticeable impact even in the cases when the
strict alignment is not required.

Change-Id: I809db3607e9e4fbd48201248355e1c9a8db99c17