boot_modules: allocate MBI structure at modaddr
This doesn't make any difference for MMU systems. But on MPU systems,
putting the MBI structure next to the modules, makes sure that the
memory map is not fragmented too much. Previously the MBI structure
would be allocated at the beginning of the RAM with no means to control
the location.

Change-Id: Ie5663a42d979b1e6a846a62a3db1691212237704