Make option and argument parsing more consistent
This makes the option parsing return the pointer to after the found
option, as is documented, instead of to the start. This is also the
way it is always used, with the callsites usually moving the pointer
immediately.

The callsites are adapted accordingly.

Also remove the realmode comment which is irrelevant for the function.

While at it incorporate a suggestion to warn of styles with arbitrary
separators between options and their arguments. We should enforce a
consistent style that is also documented, separating arguments from
options by the '=' character. In the interest of compatibility only warn
in case of a different character instead of rejecting it for now.

Change-Id: I43e3f0e9851a02583f38e462c314dde04db67008