UEFI GOP: Compare the mode information structures logically
Instead of comparing the EFI_GRAPHICS_OUTPUT_MODE_INFORMATION
structures for equality in a byte-by-byte fashion using memcmp,
compare them logically.

This eliminates false negative equality comparisons in two possible
situations:

(1) When the structures differ only in the padding bytes (e.g. due to
    some random garbage in the memory).

(2) When the structures differ only in the PixelInformation member
    (e.g. due to missing initialization) while its data is not relevant
    since the PixelFormat member does not contain the PixelBitMask
    value.

Change-Id: Ia120db69929343d05f3d8f56b42776ba2f30bd54