ARCH-amd64/libc32: re-synchronize with kernel libc
Keep the code synchronized and remove some unused structure members.

None of these changes is intended to change functionality. The fix in
puts.c addresses a problem if stdio locking is enabled (disabled here).

The function putstr() is not used and actually is no-standard so remove
its implementation. The function putchar() is currently unused but might
be used in the future so keep that one and repair it.

Change-Id: I95697f4c8753a214ef2c4d53d01d95b935790afe