[API] EFI: Improve RSDP/XSDP pointer handling
On EFI systems, the ACPI RSDP/XSDP pointer does not need to be searched,
the value given by EFI is the right value. Further, the search logic for
the region assumes 0x10 increments, whereas on EFI the pointer can have
any (pointer-aligned) address, thus the search logic does not work.
Further, the Kip regions are 0x400 aligned, which does not allow to
expess the pointers actual value.
Thus, to make the RSDP/XSDP address known to the kernel and io, we use a
field in the KIP to transfer it.

Change-Id: I34056ac78a07439b968d984fe913dd292ee51442
Requires-Change-Id: I931f435ae8b6c97528df342ac3d5ecb2ba60039f