Reserve the RSDP region before placing modules
The RSDP table is stored among other tags by a multiboot2 boot loader.
In general, that is an arbitrary location within the available
(non-reserved) memory.

If we don't relocate the RSDP table soon enough, the original location
might be overwritten by other content or conflict with a different
region (e.g. in case of GRUB 2, the original RSDP location conflicts
with the future location of sigma0).

To avoid this, we copy the RSDP table into a dedicated memory allocated
and reserved in construct_mbi(), i.e. before the modules are placed.

While at it, we pass the RSDP size (instead of the end address) between
the methods and use more semantically appropriate types to make the code
more readable.

Change-Id: I18c263299b99d82ec80a29f6c2dfbc9f1f9e252f