Remake handler logic for elf parsing and exec
We pass through the module explicitly now as a reference, since everyone
got it anyways via the handle. Also use a void* as opaque handle and
rename it to opaque because it is completely passed through to the
handler and not interpreted by the parser anymore. To achieve that move
the entry reading to startup.cc directly.

Change-Id: Id68e0bec1dd05bc654239660379ca32a9f7794e3