Provide magic __ehdr_start symbol
This symbol is going to be used to determine the load address of a PIE
binary. It is usually provided automatically by the linker if the ELF
headers are part of a load section. For bootstrap we have to provide it
ourself for two reasons, though.

 1. The ELF headers are not needed. Thus, the linker will not create
    the magic __ehdr_start symbol.
 2. We must link bootstrap to its real address. Otherwise it will break
    the boot via elfimage probably on all platforms.

The solution is to provide __ehdr_start start as absolute symbol with
address 0.

Change-Id: I825eb46aeced6e5043b9d53fa66ef7b93f29f0a2