Creation date and sequence number for L4Images
Since images are now mostly build with L4Image instead of the
buildsystem the build number and build date of bootstrap itself are less
useful to determine when an image was built.

This change adapts bootstrap to use creation and modification
timestamps added to the L4Image global attributes. It also uses a
sequence number which can be added through the 'l4i:sequence' attribute.

For example the main L4Re Makefile uses this to implement a global build
sequence counter which is incremented with each image build through that
makefile.

While at it also make the compiler version output a bit clearer.

Change-Id: I7ba9b55ff2183579203e5f23601d420379dd859c