Remove old code for small file offset compatibility
After enabling _FILE_OFFSET_BITS=64 per default for L4Re it is no longer
needed to work around the compiler defining this as well. Instead remove
the definition of the offending lseek64 function in case it is just an
alias to the lseek function (i.e. if _FILE_OFFSET_BITS is 64).

This removes the need for the adaptations from change

Ieea60882169ccb8835f0a3e1d7b5b33e244af4b0

It also improves compatibility with other C libraries that do not have a
libc-symbols header.

Change-Id: Ief0abbed42f27d427edf6e3e9d4187bedb0b9b5c