boot32: Do not use the installed libc includes
The default CPPFLAGS include LIBCINCDIR, which points to the include
directory of the installed libc. This must not be used, because we need
the libc specific to the 32 bit architecture for the boot32 part. So
override the path to the LIBCINCDIR to only include the GCC library and
the -nostdinc flag.

Also indicate that we build for L4 to avoid pulling in unnecessary
headers and remove some unused header references. Replace the _exit
function with _Exit which is defined in stdlib.h instead of unistd.h.

Change-Id: I2a1d30e55c20bb9c356137cdfb51346040267239