mk: Use CURDIR instead of PWD
The use of $(PWD) in Makefiles is strongly discouraged, since it leads
to a confusing user experience. In particular, this is the environment
variable for the PWD that was current with the original make invocation.

So a call to `/foo/bar$ make -C /path/to/dir` will set this to /foo/bar
and not to /path/to/dir. Further sub-makes invoked per `$(MAKE) -C xyz`
will keep the original value. If it needs changing this must be done
manually.

CURDIR, in the contrary is updated with any `$(MAKE) -C ` invocation and
corresponds to the path that a rule executing just `pwd` would show.
This is more in line with user expectations.

In particular, the previous behavior did not allow to call make -C on a
directory with a subdir rule, since some of the checks were performed on
the PWD instead of the CURDIR set with -C.

Change-Id: I746c85506dcda41394609e4357676749792d54d3