Increase alignment of CTORS in linker scripts
On amd64 EFI builds I observed execution failures which raised an #UD
exception in QEMU.
Debugging lead to the execution of
  call_ctors(__CTORS_BEGIN, __CTORS_END);
in libc_support+.cc. Further analysis with nm showed the .ctors list as
containing one element, but the included file shown in `ld --print-map`
output contains no entry for said list. Thus, the .ctors list should be
empty and __CTORS_BEGIN and __CTORS_END should have the same address.

Further investigation of the `ld --print-map` output shows a
  *fill* 0x8
statement, which adds a padding to between __CTORS_BEGIN & __CTORS_END.
We believe this padding to be of the wrong size and in the wrong place,
because the pre-padding address was already 8-byte aligned and thus
satisfied the linker scripts 8-byte alignment requirement.

We do not know, why ld behaves this way. Increasing the alignment
circumvents the issue.

Change-Id: I21112b05fbdb5660de0d6dbfda7b7adce319c343